报表

+++ src/pageDetail/purchasingManage/purchasingManage.vue
@@ -32,7 +32,7 @@
           <div>
             <el-button type="text" @click="detailBtn(scope.row.id)">查看</el-button>
             <el-button type="text" @click="updateBtn(scope.row.id)" v-if="!scope.row.storage">编辑</el-button>
-             <el-button type="text" @click="deleteBtn(scope.row)" v-if="!scope.row.storage" style="color: red">删除</el-button>
+             <el-button type="text" @click="deleteBtn(scope.row.id)" v-if="!scope.row.storage" style="color: red">删除</el-button>
             <el-button type="text" @click="stoBtn(scope.row)" v-if="!scope.row.storage">入库</el-button>
           </div>
         </template>
@@ -98,9 +98,26 @@ export default {
       this.getList();
     },
     deleteBtn(row) {
-      axios.delete(purchaseDelete + "?id=" + row).then(data => {
-        this.getList();
+
+      this.$confirm('此操作将永久删除该商品, 是否继续?', '提示', {
+        confirmButtonText: '确定',
+        cancelButtonText: '取消',
+        type: 'warning'
+      }).then(() => {
+        axios.delete(purchaseDelete + "?id=" + row).then(data => {
+          this.getList();
+        })
+        this.$message({
+          type: 'success',
+          message: '删除成功!'
+        });
+      }).catch(() => {
+        this.$message({
+          type: 'info',
+          message: '已取消删除'
+        });
       });
+
     },
     getList() {
       axios.post(purchaseSelect, this.theQuery).then(data => {
