+++ src/pageDetail/shipmentManage/shipmentManage.vue
@@ -100,10 +100,26 @@
         });
       },
       deleteBtn(row) {
-        axios.delete(shipmentDelete + "?id=" + row).then(data => {
-          this.$message.success("删除成功");
-          this.getList();
+
+        this.$confirm('此操作将永久删除该商品, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
+          axios.delete(shipmentDelete + "?id=" + row).then(data => {
+            this.getList();
+          });
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });
         });
+
       },
       getList() {
         axios.post(shipmentSelect, this.theQuery).then(data => {
